+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Now generates data about electricity generation (excluding CHP at the moment)
 * go.rb can now multi-thread when operating in results only mode
 * First attempt at data about road transport and heating
+* New version of cost-emissions scatter chart
 
 #2015-07-05